:recycle: Split account routes into directory module by API surface
Pure mechanical move of server/src/route/account.rs into a directory module; no behavior change.
